add home screen layout
